simple_app/requirements.txt
aiographql-client==1.0.3
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==9.0.1
anyio==3.5.0
asgi-lifespan==1.0.1
asgiref==3.5.0
async-timeout==4.0.2
asyncio-executor==0.0.4
asyncpg==0.25.0
attrs==21.4.0
bcrypt==3.2.0
cafeteria==0.20.0
cafeteria-asyncio==0.2.1
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
databases==0.5.5
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.75.1
fastapi-jwt-auth==0.5.0
frozenlist==1.3.0
graphene==3.0
graphene-pydantic==0.3.0
graphql-core==3.1.7
graphql-relay==3.1.5
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
iniconfig==1.1.1
multidict==6.0.2
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
promise==2.3
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
PyJWT==1.7.1
pyparsing==3.0.8
pytest==7.1.1
pytest-async-sqlalchemy==0.2.0
pytest-asyncio==0.18.3
python-dotenv==0.20.0
python-gql==0.2.4
python-jose==3.3.0
PyYAML==5.4.1
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
sqlalchemy2-stubs==0.0.2a21
sqlmodel==0.0.6
starlette==0.17.1
starlette-graphene3==0.5.1
starlette-graphql==0.2.1
tomli==2.0.1
typing_extensions==4.1.1
ujson==5.2.0
urllib3==1.26.9
uvicorn==0.17.6
websockets==10.2
yarl==1.7.2
